Update - Lexer & Parser Functionality
Lexer no longer uses a lazy function to create the readable struct, and
instead requires a buffer instead.
Lexer structs are now copyable.

NOTES
* Rust fundamentals
I realized that I was trying to use my OOP principles into rust.
Like using the struct items in a list as pointers. This lead to increased
lifetime complexity and trying to claim ownership of list items.

The nice thing about rust ownership is that all "moved" variables get dropped.
So instead of trying to work with token references all the time.

The parser owns the tokens and copies them when doing postfix ordering
and converting them into an AST.

This not only removed the headache using token references, but also reduced the
lifetime complexity as well.

* rust standard output
I don't understand why, but print! only appends to the standard output buffer.
To ensure that it get's printed, you have to call `io::stdout().flush().unwrap()`.

Why that is, i have no idea. but this fixed my cli formatting issues.