Revamped parsing logic
While re-ordering the tokens into postfix order works for simple
expression parsing, it falls apart when the grammar supports other
features like function calls (e.g. foo(arg)).

postfix order sequences omit the requirement for parentheses.
So we would see the function name symbol but not the parentheses that
close the argument variables.

To move away from this, I need to change the operator priority parsing
to ensure BEDMAS is enforced. refer to the `parser.md` file for the changes.

Now the parser uses helper methods that recursively get called based on
the current token the parser is on.
This not only makes the logic consistent, but clean as well. We can use
the grammar note to implement it's equivalent code.