interpreter errors print correctly
To ensure that the interpreter works as expected.
We pass a execution context through the interpreter call stack.

This keeps track of the the current position the interpreter is processing.
As well as the source code that is being interpreted.

To do this, the parser nodes have a "column range" or "position" ((i, j) where j > i).
When the interpreter visits each node, the execution context current position
is then set to the node's.

So if there are any interpreter errors (like divide by zero), the error struct uses
the execution to set the bounds that this error occured in.