Global symbol table functioning
I revamped the execution context to use a reference of a table,
and changed the symbol table to use a RefCell for the key/value
contents.

I fell into the OOP trap of trying to have a mutable reference
throughout the recursive interpreter stacktrace (this can lead to
dangling pointers in C). Instead, I opted to only accept a reference
to a symbol table and have a mutable interior key/value store.

The limitation now is that in order to return the value of a symbol,
we can no longer use `self.symbols.get`. It has to borrow the
refcell, which disallows returning a reference from the borrowed refcell
value.

The solution is to return a copy of the symbol value. This works fine as
long as we have copyable values.
This may change when we want to hold functions in the symbol table.